Fetch profile data from links everywhere
