+++ src/containers/Profile.vue
@@ -14,7 +14,7 @@
         {{ `${profile.givenName} ${profile.familyName}` }}
       </j-text>
       <j-text variant="body"> @{{ profile.username }}</j-text>
-      <j-text v-if="bio" variant="subheading"> {{ bio }}</j-text>
+      <j-text v-if="profile.bio" variant="subheading"> {{ profile.bio }}</j-text>
       <j-button @click="() => $emit('openCompleteProfile')">
         View complete profile
       </j-button>
@@ -30,11 +30,9 @@
 
 <script lang="ts">
 import { defineComponent } from "vue";
-import { getProfile, parseProfile } from "@/utils/profileHelpers";
 import { Profile } from "@/store/types";
 import Skeleton from "@/components/skeleton/Skeleton.vue";
-import { LinkExpression, Perspective } from "@perspect3vism/ad4m";
-import { ad4mClient } from "@/app";
+import { getProfile } from "@/utils/profileHelpers";
 
 export default defineComponent({
   components: { Skeleton },
@@ -43,8 +41,6 @@ export default defineComponent({
   data() {
     return {
       profile: null as null | Profile,
-      agentPerspective: undefined as undefined | Perspective | null,
-      bio: "",
     };
   },
   watch: {
@@ -52,21 +48,7 @@ export default defineComponent({
       handler: async function (did) {
         // reset profile before fetching again
         this.profile = null;
-        let profileLang = this.langAddress;
-        const dataExp = await getProfile(profileLang, did);
-        if (dataExp) {
-          this.profile = dataExp;
-        }
-
-        const profilePerspective = await ad4mClient.agent.byDID(did);
-        this.agentPerspective = profilePerspective.perspective;
-
-        const bioLink = profilePerspective.perspective?.links.find(
-          (e) => e.data.predicate === "sioc://has_bio"
-        ) as LinkExpression;
-        if (bioLink) {
-          this.bio = bioLink.data.target.split("://")[1];
-        }
+        this.profile = await getProfile(did);
       },
       immediate: true,
     },
